Attempted to support translation as part of ES2 but hit some problems that I don't have time to sort out right now.  Just set the IOSImplementation.isTranslationSupported() to return false for now.
Added support for 3d transforms by way of the com.codename1.ui.geom.Matrix class and the Graphics.setTransform(Matrix), and Graphics.setTransform(Matrix, originX, originY) methods.

Fixed The rasterization of shapes to not be affected by clipping.  This required an extra call to getBounds() which might end up being an expense we want to minimize, but it was necessary.

At this point it appears to work pretty well with ES2 - but haven't tried ES1 yet for regressions.  Also, still need to add support for drawing alpha masks and texture caching for supporting hi performance animations.